fix_db_2.0

+++ server/db/models/favourite.js
@@ -8,8 +8,8 @@ module.exports = (sequelize, DataTypes) => {
     }
   }
   Favourite.init({
-    userId: DataTypes.BIGINT,
-    sockId: DataTypes.BIGINT
+    userId: DataTypes.INTEGER,
+    sockId: DataTypes.INTEGER
   }, {
     sequelize,
     modelName: 'Favourite',
